+++ CHANGELOG.md
@@ -8,6 +8,8 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - removed 6 unneeded fields from school json payload
 - Added notifications script and 2 fields to Notification model for errors
 - Made program code required/program name not required in load programs script
+- Made school name bold
+- Use total direct cost value from offer URL
 - Update toggle functionality, interactions, and dynamic content
 
 ## 2.1.2